added labs for lecture #2

+++ src/proc/recursion/hanoi_tower/Hanoi.java
@@ -1,6 +1,5 @@
 package proc.recursion.hanoi_tower;
 
-import java.util.Arrays;
 import java.util.Stack;
 
 /**
@@ -29,7 +28,7 @@ public class Hanoi {
             Stack<Integer> to, int count) {
         if (count > 0) {
             exchange(from, to, help, count - 1);
-            int biggest = from.pop();
+            int biggest = (int)from.pop(); // cast to an int
             to.push(biggest);
             exchange(help, from, to, count - 1);
         }
